% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuromorphDatasets.R
\docType{data}
\name{cell.type.diff}
\alias{cell.type.diff}
\title{Diferentes tipos celulares en Rata}
\format{
A data frame with 38559 rows and 10 variables:
\describe{
\item{species}{Especies de los organismos registrados}
\item{brain_region}{Region a la que pertenece el registro}
\item{cell_type}{Tipo de celula que se registra}
\item{domain}{Dominio estructural de la célula}
\item{soma_surface}{Superficie del soma en \eqn{\mum^{2}}}
\item{surface}{Superficie de la célula en \eqn{\mum^{2}}}
\item{volume}{Superficie de la célula en \eqn{\mum^{3}}}
\item{sex}{Sexo del organismo}
\item{age_classification}{Etiqueta de edad del organismo}
\item{min_age}{Edad mínima en meses del rango definido}
\item{max_age}{Edad máxima en meses del rango definido}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
cell.type.diff
}
\description{
Conjunto de datos obtenidos de \href{http://neuromorpho.org/}{Neuromorpho} mediante el paquete
\href{https://github.com/natverse/neuromorphr}{neuromorphr}. Este dataset contiene información
de diferentes tipo celulares de rata.
}
\keyword{datasets}
