% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNATBIDataSets.R
\docType{data}
\name{column.samples}
\alias{column.samples}
\title{Información de las muestras}
\format{
Data frame con 377 renglones y 12 columnas:
\describe{
\item{rnaseq_profile_id}{Identificador de sequencia de RNA}
\item{donor_id}{Idenficador de donante}
\item{donor_name}{Etiqueta del donante}
\item{specimen_id}{Identificador de muestra}
\item{specime_name}{Etiqueta de la muestra}
\item{rna_well_id}{Identificador del pozo}
\item{polygon_id}{Identificador del polígono}
\item{structure_id}{Identificaodr de estructura}
\item{structure_acronym}{Acronimo de la estructura}
\item{structure_color}{Color de la estructura}
\item{structure_name}{Nombre de la estructura}
\item{hemisphere}{Hemisferio}
}
}
\source{
\url{http://aging.brain-map.org/}
}
\usage{
column.samples
}
\description{
En esta tabla encontramos la información de las muestras tomadas, identificador
del donador, nombre de la estructura de donde se toma y hemisferio correspondiente
}
\keyword{datasets}
