% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuromorphDatasets.R
\docType{data}
\name{donor.information}
\alias{donor.information}
\title{Información del donante}
\format{
A data frame with 107 rows and 19 variables:
\describe{
\item{donor_id}{Número de donador}
\item{name}{Identificador del donador anonimizado}
\item{age}{Edad}
\item{sex}{Sexo}
\item{apo_e4_allele}{Yes - si presenta un ApoE4 alelo; No - sin ApoE4 alelos presentes}
\item{education_years}{Número de años de educación completada}
\item{age_at_first_tbi}{Edad en la que se reporta el primer TBI con perdida de memoria.}
\item{longest_loc_duration}{Puntuación del tiempo de perdida de conciencia: 0 (no o desconocido), 1 (segundoso menos), 2 (minutos o menos), 3 (1-2 min), 4 (3-5 min), 5 (6-9 min), 6 (10 min-1 hr), y 7 (> 1 hr).}
\item{cerad}{Estado de la severidad de disposición amiloide de 0 (ninguna) a 3 (severa), CERAD  (Consortium to Establish a Registry for Alzheimer’s Disease)}
\item{num_tbi_w_loc}{Numbero de TBI registradosdesde 0 (control) a 3.}
\item{dsm_iv_clinical_diagnosis}{Classificación encontrada en "Diagnostic and Statistical Manual of Mental Disorders, 4th Edition", para la presencia de demencia subtipo 1: 1 (Alzheimer's), 2 (vascular), 3 (otras médicas), 4 (sustancia), 5 (multiple), 6 (otras/desconocido), y 7 (no demencia).}
\item{control_set}{Entero único correspondiente a cada par de TBI y donadores control pareados.}
\item{nincds_arda_diagnosis}{Descripción clínica de probable o posible Alzheimer's basado en criterios NINCDS/ADRDA: 0 (sin demencia), 1 (probable Alzheimer's), 2 (posible Alzheimer's disease), y 3 (dementia, tipo desconocido).}
\item{ever_tbi_w_loc}{Identificación de TBI vs. casos control}
\item{race}{Blanco o no blanco}
\item{hispanic}{Hispano o no hispano}
\item{act_demented}{Cualquier demencia detectada durante el estudio ACT.}
\item{braak}{Andamiaje de maraña neurofibrillary (NFTs) de 0 (none) a 6 (NFTs neocorticales extendidos) basados en el método Braak.}
\item{nia_reagan}{Probabilida de demencia clínica como resultado de la enfermedad Alzheimer's.}
}
}
\source{
\url{http://aging.brain-map.org/}
}
\usage{
donor.information
}
\description{
Información clínica anonimizada que incluye Alzheimer, demencia y diagnostico  de TBI (Traumatic Brain Injury)
para los donantes en el estudio. Estos datos son proporcionados por:
© 2016 Allen Institute for Cell Science. Aging, Dementia and TBI. Available from: \href{http://aging.brain-map.org/}{http://aging.brain-map.org/}
}
\details{
Mas información sobre la base de datos en la
\href{https://help.brain-map.org/display/aging/Documentation}{Documentación}
}
\keyword{datasets}
