% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseRNATBIData.R
\name{GetRNADataSet}
\alias{GetRNADataSet}
\title{Get RNA dataset}
\usage{
GetRNADataSet(save.file = TRUE, clean.tmp.files = FALSE)
}
\arguments{
\item{save.file}{Parámetro booleano que define si se guardar la tabla procesada}

\item{clean.tmp.files}{Parámetro booleano que si es verdadero borrará el archivo
con los datos originales}
}
\value{
data.table/dataframe con la información de la expresión de RNA
}
\description{
Esta función permite descargar y procesar la información de RNA de TBI
usa data.table para descargarla y optimizar su uso en memoria, pero el resultado
es compatible con la sintaxis dataframe.
Ademas permite definir si se quiere guardar el set de datos que se procesa a
partir de los archivos originales.
Otro parámetro a considerar es clean.tmp.files que permite borrar ó conservar
el archivo original con todos los datos.
}
\details{
Datos con las tomas de expresión normalizadas para los diferentes genes
y los diferentes perfiles de las sequencias de RNA.

Para saber a que gen pertenece cada muestra es necesario cruzar los datos con
los datos de rows.genes, y para las sequecias de RNA el cruce debe hacerse con
los datos de información de las muestras.

\itemize{
\item{gene_id}{Número de identificación del gen}
\item{rnaseq_profile_id}{Identificador de sequencia de RNA}
\item{value}{Valores normalizados de la expresión del RNA}
}
}
\examples{
\dontrun{
  GetRNADataSet(save.file = TRUE, clean.tmp.files = FALSE )
}
}
