% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuromorphDatasets.R
\docType{data}
\name{neocortex.animal.size}
\alias{neocortex.animal.size}
\title{Neocorteza y tamaño animal}
\format{
Data frame con 24709 renglones y 11 columnas:
\describe{
\item{species}{Especies de los organismos registrados}
\item{brain_region}{Region a la que pertenece el registro}
\item{cell_type}{Tipo de celula que se registra}
\item{domain}{Dominio estructural de la célula}
\item{soma_surface}{Superficie del soma en \eqn{\mum^{2}}}
\item{surface}{Superficie de la célula en \eqn{\mum^{2}}}
\item{volume}{Superficie de la célula en \eqn{\mum^{3}}}
\item{sex}{Sexo del organismo}
\item{age_classification}{Etiqueta de edad del organismo}
\item{min_age}{Edad mínima en meses del rango definido}
\item{max_age}{Edad máxima en meses del rango definido}
}
}
\source{
\url{http://neuromorpho.org/}
}
\usage{
neocortex.animal.size
}
\description{
Conjunto de datos obtenidos de \href{http://neuromorpho.org/}{Neuromorpho} mediante el paquete
\href{https://github.com/natverse/neuromorphr}{neuromorphr}.
Estos datos contienen 24709 registros de características de
células de la neocorteza para tres especies de animales rata, elefante y
ballena jorobada.
}
\keyword{datasets}
