% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeurocienciasIntroData-package.R
\docType{package}
\name{NeurocienciasIntroData-package}
\alias{NeurocienciasIntroData}
\alias{NeurocienciasIntroData-package}
\title{NeurocienciasIntroData: Paquete con Datos y Tutorales de Análisis de Datos}
\description{
Paquete con datos para realizar exploración de datos y proyectos del
  curso.Incluye datos morfológicos de células  tanto de  superficie celular, 
  volumen y superficie de soma, entre otras características recabandas para 
  celulas del sistema nervioso de diversos organismos.
  Estos datos se obtienen de la página neuromorpho.org usando el paquete neuromorphr.
  Incluye : cell.type.diff, principal.cell.diff y neocortex.animal.size
  Tambien incluye datos de Aging, Dementia and Traumatic Brain Injury Study,
  donde se presentan datos de expresión de proteínas y mRNA relacionado
  con TBI.
  Incluye: donor.information, protein.pathology, rows.genes y column.samples
  Tambien incluye un fragmento de los datos proporcionados por el gobierno 
  de México sobre los casos de COVID.
  Incluye: datos.abiertos.covid y reporte.casos.covid.
  Para mas información se pueden consultar la sección de ayuda para cada uno de
  estos set de datos.
}
\author{
\strong{Maintainer}: Edgar Morales \email{edgar.morales@aprendr.org}

}
\keyword{internal}
