% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNATBIDataSets.R
\docType{data}
\name{normalized.rna.measures}
\alias{normalized.rna.measures}
\title{Expresión de RNA normalizada}
\format{
Data frame con 18955937 renglones y 3 columnas:
\describe{
\item{gene_id}{Número de identificación del gen}
\item{rnaseq_profile_id}{Identificador de sequencia de RNA}
\item{value}{Valores normalizados de la expresión del RNA}
}
}
\source{
\url{http://aging.brain-map.org/}
}
\usage{
normalized.rna.measures
}
\description{
Datos con las tomas de expresión normalizadas para los diferentes genes
y los diferentes perfiles de las sequencias de RNA.
}
\details{
Para saber a que gen pertenece cada muestra es necesario cruzar los datos con
los datos de rows.genes, y para las sequecias de RNA el cruce debe hacerse con
los datos de información de las muestras.
}
\keyword{datasets}
