% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuromorphDatasets.R
\docType{data}
\name{reporte.casos.covid}
\alias{reporte.casos.covid}
\title{Reporte de casos covid}
\format{
Data frame con 70 renglones y 20 columnas:
\describe{
\item{Fecha}{ Fecha del Registro}
\item{SospechososNuevos}{Nuevos casos sospechosos}
\item{SospechososAcum}{Casos sospechados acumulados}
\item{SospechososPorcenCambio}{Porcentaje de cambio de casos sospechosos}
\item{NegativosNuevos}{Casos nuevos negativos}
\item{NegativosAcum}{Casos Negativos Acumulados}
\item{NegativosPorcenCambio}{Porcentaje de cambio casos acumulados}
\item{ConfirmadosNuevos}{Casos confirmados nuevos}
\item{ConfirmadosAcum}{Casos confirmados acumulados}
\item{ConfirmadosPorcenCambio}{Porcentaje de cambio confirmados}
\item{RecuperadosNuevos}{Casos recuperados nuevos}
\item{RecuperadosAcum}{Acumulado de casos recuperados}
\item{RecuperadosPorcenCambio}{Porcentaje de cambio de casos recuperados}
\item{FallecidosNuevos}{Fallecimientos nuevos}
\item{FallecidosAcum}{Fallecimientos acumulados}
\item{FallecidosPorcenCambio}{Porcentaje de cambio fallecidos}
\item{EstudiadosNuevos}{Estudios nuevos}
\item{EstudiadosAcum}{Estudios acumulados}
\item{EstudiadosPorcenCambio}{Porcentaje de cambio de estudios}
\item{ActivosAcum}{Activos acumulados}
}
}
\source{
\url{https://www.gob.mx/salud/es/archivo/documentos}
}
\usage{
reporte.casos.covid
}
\description{
Resumen de casos covid reportados por el Gobierno de México la totalidad de estos datos se
pueden consultar en la página de la \href{https://www.gob.mx/salud/es/archivo/documentos}{secretaría de salud}.
}
\keyword{datasets}
